Fixes wrong default value behaviour in Event.time_left
In python the default value of an argument gets cached the first
time the function is invoked. Subsequent calls to the function
without specifying the argument will use the same cached default
value again. In this particular case all relative deltas will be
calculated from the now that was cached when the function was
invoked the first time thus leading to a wrong time_left